Refactor compiler into multiple source files
