+++ src/parser.c
@@ -1,0 +1,160 @@
+#include "dream.h"
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+
+Node *create_node(NodeType type, char *value, Node *left, Node *right)
+{
+    Node *node = malloc(sizeof(Node));
+    node->type = type;
+    node->value = value ? strdup(value) : NULL;
+    node->left = left;
+    node->right = right;
+    return node;
+}
+
+Node *parse_expression(Lexer *lexer, Token *token)
+{
+    if (token->type != TOKEN_IDENTIFIER && token->type != TOKEN_NUMBER) {
+        fprintf(stderr, "Expected identifier or number\n");
+        exit(1);
+    }
+    Node *left = create_node(token->type == TOKEN_IDENTIFIER ? NODE_IDENTIFIER : NODE_NUMBER,
+                             token->value, NULL, NULL);
+    *token = next_token(lexer);
+    if (token->type == TOKEN_PLUS) {
+        char *op = token->value;
+        *token = next_token(lexer);
+        if (token->type != TOKEN_IDENTIFIER && token->type != TOKEN_NUMBER) {
+            fprintf(stderr, "Expected identifier or number after +\n");
+            exit(1);
+        }
+        Node *right = create_node(token->type == TOKEN_IDENTIFIER ? NODE_IDENTIFIER : NODE_NUMBER,
+                                  token->value, NULL, NULL);
+        *token = next_token(lexer);
+        return create_node(NODE_BINARY_OP, op, left, right);
+    }
+    return left;
+}
+
+Node *parse_statement(Lexer *lexer, Token *token)
+{
+    if (token->type == TOKEN_INT) {
+        free(token->value);
+        *token = next_token(lexer);
+        if (token->type != TOKEN_IDENTIFIER) {
+            fprintf(stderr, "Expected identifier after int\n");
+            exit(1);
+        }
+        char *var_name = token->value;
+        *token = next_token(lexer);
+        Node *init = NULL;
+        if (token->type == TOKEN_EQUALS) {
+            free(token->value);
+            *token = next_token(lexer);
+            init = parse_expression(lexer, token);
+        }
+        if (token->type != TOKEN_SEMICOLON) {
+            fprintf(stderr, "Expected semicolon\n");
+            exit(1);
+        }
+        return create_node(NODE_VAR_DECL, var_name, init, NULL);
+    } else if (token->type == TOKEN_IDENTIFIER) {
+        char *var_name = token->value;
+        *token = next_token(lexer);
+        if (token->type == TOKEN_EQUALS) {
+            free(token->value);
+            *token = next_token(lexer);
+            Node *expr = parse_expression(lexer, token);
+            if (token->type != TOKEN_SEMICOLON) {
+                fprintf(stderr, "Expected semicolon\n");
+                exit(1);
+            }
+            return create_node(NODE_ASSIGN, var_name, expr, NULL);
+        }
+    } else if (token->type == TOKEN_CONSOLE) {
+        free(token->value);
+        *token = next_token(lexer);
+        if (token->type != TOKEN_DOT) {
+            fprintf(stderr, "Expected dot after Console\n");
+            exit(1);
+        }
+        free(token->value);
+        *token = next_token(lexer);
+        if (token->type != TOKEN_WRITELINE) {
+            fprintf(stderr, "Expected WriteLine\n");
+            exit(1);
+        }
+        free(token->value);
+        *token = next_token(lexer);
+        if (token->type != TOKEN_LPAREN) {
+            fprintf(stderr, "Expected (\n");
+            exit(1);
+        }
+        free(token->value);
+        *token = next_token(lexer);
+        if (token->type != TOKEN_IDENTIFIER && token->type != TOKEN_NUMBER) {
+            fprintf(stderr, "Expected identifier or number\n");
+            exit(1);
+        }
+        Node *arg = create_node(token->type == TOKEN_IDENTIFIER ? NODE_IDENTIFIER : NODE_NUMBER,
+                               token->value, NULL, NULL);
+        *token = next_token(lexer);
+        if (token->type != TOKEN_RPAREN) {
+            fprintf(stderr, "Expected )\n");
+            exit(1);
+        }
+        free(token->value);
+        *token = next_token(lexer);
+        if (token->type != TOKEN_SEMICOLON) {
+            fprintf(stderr, "Expected semicolon\n");
+            exit(1);
+        }
+        return create_node(NODE_WRITELINE, NULL, arg, NULL);
+    } else if (token->type == TOKEN_IF) {
+        free(token->value);
+        *token = next_token(lexer);
+        if (token->type != TOKEN_LPAREN) {
+            fprintf(stderr, "Expected ( after if\n");
+            exit(1);
+        }
+        free(token->value);
+        *token = next_token(lexer);
+        Node *cond = parse_expression(lexer, token);
+        if (token->type != TOKEN_RPAREN) {
+            fprintf(stderr, "Expected ) after condition\n");
+            exit(1);
+        }
+        free(token->value);
+        *token = next_token(lexer);
+        Node *body = NULL;
+        if (token->type == TOKEN_LBRACE) {
+            free(token->value);
+            *token = next_token(lexer);
+            body = parse_statement(lexer, token);
+            if (token->type != TOKEN_RBRACE) {
+                fprintf(stderr, "Expected } after if body\n");
+                exit(1);
+            }
+            free(token->value);
+            *token = next_token(lexer);
+        } else {
+            body = parse_statement(lexer, token);
+        }
+        return create_node(NODE_IF, NULL, cond, body);
+    }
+    fprintf(stderr, "Invalid statement\n");
+    exit(1);
+}
+
+void free_node(Node *n)
+{
+    if (!n)
+        return;
+    if (n->value)
+        free(n->value);
+    free_node(n->left);
+    free_node(n->right);
+    free(n);
+}
+
